GVT-1551: Split publication-table into separate preview and publication tables
